[test-app] Unbreak MultipleWebviewHandlingTests
The test previously hardcoded the port 8080. This works when you run a
test in isolation, because the odds of getting that port on the device
are pretty high. If multiple tests are run after each other the driver
will end up handing out higher ports. For now I fixed it by changing the
port to the port of the selendroid standalone driver that is used in all
these tests (4444). This works because the selendroid standalone driver
will forward requests to unknown paths to the device by default.