[server] Fix ClassPredicate
While looking into the failures of `NativeChildElementFindingTest` I
found out that the `ClassPredicate` is throwing an exception in case the
class is not found in the app. This regressed in
4afa5ccedc53b0b73319dfe3a6917db0f557fddc when the class predicate was
extracted to its own class.

I changed the class predicate to return false if the class is not found,
but logging a warning for discoverability. I changed the tests to use an
existing class in the normal case and added new tests to guard for the
appropriate behavior if the class does not exist.