[test-app] Unbreak NativeElementFindingTest
Most "broken" tests weren't updated for UI changes in the app. Either
asserting the incorrect amount of elements, or wrong text string. I
fixed those cases.

The `shouldNotFindElementByIdFromPreviousActivity()` test not broken,
but exposes a bug in selendroid where selendroid will look for elements
in multiple top level views. The bug is is tracked in
https://github.com/selendroid/selendroid/issues/800. I am disabling the
test until the issue is fixed.